+++ src/voice/voice_liveness_detector.cpp
@@ -294,8 +294,8 @@ std::string VoiceLivenessDetector::generateRandomChallenge() {
     static std::random_device rd;
     static std::mt19937 gen(rd());
 
-    std::uniform_int_distribution<size_t> dis(0, CHALLENGE_PHRASES.size() - 1);
-    return CHALLENGE_PHRASES[dis(gen)];
+    std::uniform_int_distribution<uint64_t> dis(0, CHALLENGE_PHRASES.size() - 1);
+    return CHALLENGE_PHRASES[static_cast<size_t>(dis(gen))];
 }
 
 int64_t VoiceLivenessDetector::nowMs() const {
